refactor(recipe-layouts): migrate to app dir
