+++ app/(browse)/browse/components/recipe-layouts/recipe-duration.tsx
@@ -2,7 +2,7 @@ import clsx from "clsx";
 import { startCase } from "lodash";
 import React from "react";
 
-const LayoutDuration: React.FunctionComponent<{ label: string; data?: string | number }> = ({ label, data }) => {
+const RecipeDuration: React.FunctionComponent<{ label: string; data?: string | number }> = ({ label, data }) => {
   return (
     <div className={clsx("flex flex-col flex-wrap gap-x-2 pr-1 md:flex-row", !data && "hidden")}>
       <dt className="text-sm font-medium uppercase text-neutral-500">{startCase(label)}</dt>
@@ -11,4 +11,4 @@ const LayoutDuration: React.FunctionComponent<{ label: string; data?: string | n
   );
 };
 
-export default LayoutDuration;
+export default RecipeDuration;
