--- components/home/index.ts
@@ -1,2 +1,0 @@
-export { default as HomepageInfo } from "./homepage-info"
-export { default as HomepageRecipe } from "./homepage-recipe"
