--- components/home/types.d.ts
@@ -1,13 +1,0 @@
-export type QueryParam = {
-  url: string
-}
-
-export type Stats = {
-  starGazer?: number
-  origin?: string
-  forkUrl: string
-}
-
-export type HomePageProps = {
-  stats: Stats
-}
