--- components/ui/recipe/index.ts
@@ -1,3 +1,0 @@
-export { default as RecipeLayout } from "./Recipe-layout"
-export { default as RecipeLayoutSkeleton } from "./Recipe-layout-skeleton"
-export { default as RecipeUndefined } from "./Recipe-undefined"
